[room] Fix bug: don't clear event listeners after a .disconnect() call
It means that people would no longer get any notifications after calling
`.disconnect()` and `.connect()` on the same room instance. Now, this
isn't a real usage pattern that many people will ever do in reality.
More likely is that they will call `.leave()` and `.enter()` on the
Client instead. Doing so is fine, because it will destroy and recreate
the Room instance.