Simplify req/res handling
It now ignores received server messages for a request ID that already
timed out, or which we don't recognize as ours.

It won't trigger side effects for messages it did not recognize.

Also it simplifies/unifies the many, very similar, TypeScript types, by
directly letting each promise resolve to the literal network response
message, without stripping or mutating fields.

This saves a lot of translating between layers for now, while we're
still iterating.